pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "object-paint-agent"
version = "0.1.0"
description = "Object Paint Agent: detect, segment, and recolor objects in images (CPU-friendly Gradio demo)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Object Paint Agent" }]
dependencies = [
    "gradio>=4.44.0",
    "numpy>=1.24.0,<2",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
object-paint-agent = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["app", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
